# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * MediaHub API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiError {
  error: string;
}

export interface MediaInfoInput {
  /** YouTube or Instagram URL to fetch media info for */
  url: string;
}

export interface VideoFormat {
  formatId: string;
  /** Quality label e.g. 1080p, 720p */
  quality: string;
  /** File extension e.g. mp4, webm */
  ext: string;
  /**
     * Estimated file size in bytes
     * @nullable
     */
  filesize?: number | null;
  /**
     * Frames per second
     * @nullable
     */
  fps?: number | null;
  /** Direct download URL (pre-signed or proxied) */
  url: string;
  /** @nullable */
  vcodec?: string | null;
}

export interface AudioFormat {
  formatId: string;
  /** e.g. MP3 320kbps, MP3 128kbps, Original */
  label: string;
  ext: string;
  /** @nullable */
  filesize?: number | null;
  /** Proxied download URL */
  url: string;
  /**
     * Audio bitrate in kbps
     * @nullable
     */
  abr?: number | null;
}

export interface MediaInfo {
  title: string;
  /** Thumbnail URL */
  thumbnail: string;
  /** Duration in seconds */
  duration: number;
  /** Channel or creator name */
  channel: string;
  /**
     * Upload date string (YYYYMMDD or formatted)
     * @nullable
     */
  uploadDate: string | null;
  /** Detected platform (youtube, instagram) */
  platform: string;
  videoFormats: VideoFormat[];
  audioFormats: AudioFormat[];
}