# lib/api-zod/src/generated/types/audioFormat.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * MediaHub API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AudioFormat {
  formatId: string;
  /** e.g. MP3 320kbps, MP3 128kbps, Original */
  label: string;
  ext: string;
  /** @nullable */
  filesize?: number | null;
  /** Proxied download URL */
  url: string;
  /**
     * Audio bitrate in kbps
     * @nullable
     */
  abr?: number | null;
}