# lib/api-zod/src/generated/types/mediaInfo.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * MediaHub API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AudioFormat } from './audioFormat';
import type { VideoFormat } from './videoFormat';

export interface MediaInfo {
  title: string;
  /** Thumbnail URL */
  thumbnail: string;
  /** Duration in seconds */
  duration: number;
  /** Channel or creator name */
  channel: string;
  /**
     * Upload date string (YYYYMMDD or formatted)
     * @nullable
     */
  uploadDate: string | null;
  /** Detected platform (youtube, instagram) */
  platform: string;
  videoFormats: VideoFormat[];
  audioFormats: AudioFormat[];
}