# lib/api-zod/src/generated/types/mediaInfoInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * MediaHub API specification
 * OpenAPI spec version: 0.1.0
 */

export interface MediaInfoInput {
  /** YouTube or Instagram URL to fetch media info for */
  url: string;
}