# lib/api-zod/src/generated/types/videoFormat.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * MediaHub API specification
 * OpenAPI spec version: 0.1.0
 */

export interface VideoFormat {
  formatId: string;
  /** Quality label e.g. 1080p, 720p */
  quality: string;
  /** File extension e.g. mp4, webm */
  ext: string;
  /**
     * Estimated file size in bytes
     * @nullable
     */
  filesize?: number | null;
  /**
     * Frames per second
     * @nullable
     */
  fps?: number | null;
  /** Direct download URL (pre-signed or proxied) */
  url: string;
  /** @nullable */
  vcodec?: string | null;
}